pb back : all users can delete post/comment

+++ back/controllers/postControl.js
@@ -30,11 +30,11 @@ exports.getAllPosts = (req, res) => {
       include: [
         [
           sequelize.fn(
-            'DATE_FORMAT',
-            sequelize.col('post.createdAt'),
-            '%d/%m/%y %H:%i'
+            "DATE_FORMAT",
+            sequelize.col("post.createdAt"),
+            "%d/%m/%y %H:%i"
           ),
-          'createdAt',
+          "createdAt",
         ],
       ],
     },
@@ -47,22 +47,19 @@ exports.getAllPosts = (req, res) => {
       {
         model: Comment,
         as: "Comment",
-        include: [
-          { model: User,
-            attributes: ["username"]
-          }],
-          attributes: {
-            include: [
-              [
-                sequelize.fn(
-                  'DATE_FORMAT',
-                  sequelize.col('Comment.createdAt'),
-                  '%d/%m/%y %H:%i'
-                ),
-                'createdAt',
-              ],
+        include: [{ model: User, attributes: ["username"] }],
+        attributes: {
+          include: [
+            [
+              sequelize.fn(
+                "DATE_FORMAT",
+                sequelize.col("Comment.createdAt"),
+                "%d/%m/%y %H:%i"
+              ),
+              "createdAt",
             ],
-          },
+          ],
+        },
       },
     ],
 
@@ -100,6 +97,18 @@ exports.getOnePost = (req, res) => {
 exports.deletePost = (req, res) => {
   Post.findOne({ where: { id: req.params.id } })
     .then((Post) => {
+      // console.log("1 : " + Post.userId);
+      // console.log("2 : " + req.params.userId);
+
+      if (!Post) {
+        return res.status(404).json({ error: "Post non trouvé !" });
+      }
+
+      //soucis ici
+      if (Post.userId !== req.params.userId) {
+        return res.status(403).json({ error: "Requête non authorisée !" });
+      }
+
       if (Post.image != null) {
         const filename = Post.image.split("/images/")[1];
         fs.unlink(`images/${filename}`, (err) => {
